backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
cryptography==46.0.2
distlib==0.4.0
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==7.1.0
drf-yasg==1.21.11
filelock==3.20.0
gunicorn==21.2.0
identify==2.6.15
idna==3.10
inflection==0.5.1
nodeenv==1.9.1
packaging==25.0
platformdirs==4.5.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pywin32; platform_system == "Windows"
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.3
whitenoise==6.6.0
